requirements/base.txt
# Django
django==4.2.13

#Rest framework
djangorestframework==3.15.1

# dotenv
python-dotenv==0.19.2

#mysqlclient
mysqlclient==2.1.0

#celery
celery

#celery-beat
django-celery-beat

#celery results
django-celery-results

#redis
redis==4.1.0

#django extensions
django-extensions==3.1.5

# Pahomqtt
paho-mqtt==1.6.1  

# Pillow
Pillow

# Requests
requests